Abort cleanly if the stack isn't clean.
If the user tries to call an operation without enough arguments
we caught that.  (e.g. '3 +')

What we didn't catch, until now, was that the stack still had
too many items upon it when we came to exit.  We assume there
will be one item on the stack and we pop it off, then return
to exit.  If the stack has too many entries upon it then
our return-address will be bogus, and we'll get a segfault.

Catch the stack being too full, and exit cleanly in that case.

Added test too.